Create login view
The user is queried first and stored in a variable for later use
validates and checks login data for matches
Stores user's data in a session(dict) that made available on subsequent
requests. At the beggining of each request, if a user is logged in,
their information should be loaded and made available to other views.