Require authentification in other views
Creating, editing and deleting blog posts will require a user to be
logged in. A decorator returns a new view function that wraps the
original view it's applied to. The new function checks if a user is
loaded and redirects to the login page otherwise.If a user is loaded the
original view is called and continues normally.